programs/emailer/requirements.txt
Flask==2.3.2
flask_cors==4.0.0
python-dotenv==1.0.0
